dataengine/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
addict=2.4.0=pypi_0
annotated-types=0.7.0=pypi_0
anyio=4.4.0=pypi_0
argon2-cffi=23.1.0=pypi_0
argon2-cffi-bindings=21.2.0=pypi_0
arrow=1.3.0=pypi_0
asttokens=2.4.1=pypi_0
async-lru=2.0.4=pypi_0
attrs=23.2.0=pypi_0
babel=2.15.0=pypi_0
beautifulsoup4=4.12.3=pypi_0
blas=1.0=mkl
bleach=6.1.0=pypi_0
blinker=1.8.2=pypi_0
brotli-python=1.0.9=py311h6a678d5_8
bzip2=1.0.8=h5eee18b_6
ca-certificates=2024.3.11=h06a4308_0
cachetools=5.3.3=pypi_0
certifi=2023.7.22=pypi_0
cffi=1.16.0=pypi_0
chardet=4.0.0=pypi_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.1.7=pypi_0
cmake=3.25.0=pypi_0
comm=0.2.2=pypi_0
configargparse=1.7=pypi_0
contourpy=1.2.1=pypi_0
cryptography=43.0.1=pypi_0
cuda-cudart=11.8.89=0
cuda-cupti=11.8.87=0
cuda-libraries=11.8.0=0
cuda-nvrtc=11.8.89=0
cuda-nvtx=11.8.86=0
cuda-runtime=11.8.0=0
cuda-version=12.5=3
cycler=0.10.0=pypi_0
cython=3.0.10=pypi_0
dash=2.17.0=pypi_0
dash-core-components=2.0.0=pypi_0
dash-html-components=2.0.0=pypi_0
dash-table=5.0.0=pypi_0
debugpy=1.8.1=pypi_0
decorator=5.1.1=pypi_0
defusedxml=0.7.1=pypi_0
docstring-parser=0.16=pypi_0
executing=2.0.1=pypi_0
fastjsonschema=2.19.1=pypi_0
ffmpeg=4.3=hf484d3e_0
filelock=3.13.1=py311h06a4308_0
flask=3.0.3=pypi_0
fonttools=4.53.0=pypi_0
fqdn=1.5.1=pypi_0
freetype=2.12.1=h4a9f257_0
fsspec=2024.5.0=pypi_0
ftfy=6.2.0=pypi_0
fvcore=0.1.5.post20221221=pypi_0
gdown=5.2.0=pypi_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py311hc9b5ff0_0
gnutls=3.6.15=he1e5248_0
google-ai-generativelanguage=0.6.6=pypi_0
google-api-core=2.19.0=pypi_0
google-api-python-client=2.144.0=pypi_0
google-auth=2.29.0=pypi_0
google-auth-httplib2=0.2.0=pypi_0
google-cloud-aiplatform=1.53.0=pypi_0
google-cloud-bigquery=3.24.0=pypi_0
google-cloud-core=2.4.1=pypi_0
google-cloud-resource-manager=1.12.3=pypi_0
google-cloud-storage=2.16.0=pypi_0
google-crc32c=1.5.0=pypi_0
google-generativeai=0.7.2=pypi_0
google-resumable-media=2.7.0=pypi_0
googleapis-common-protos=1.63.1=pypi_0
gputil=1.4.0=pypi_0
groundingdino=0.1.0=pypi_0
grpc-google-iam-v1=0.13.0=pypi_0
grpcio=1.64.1=pypi_0
grpcio-status=1.62.2=pypi_0
h11=0.14.0=pypi_0
h5py=3.11.0=pypi_0
htmlmin=0.1.12=pypi_0
httpcore=1.0.5=pypi_0
httplib2=0.22.0=pypi_0
httpx=0.27.0=pypi_0
huggingface-hub=0.23.2=pypi_0
idna=2.10=pypi_0
imagehash=4.3.1=pypi_0
importlib-metadata=7.1.0=pypi_0
intel-openmp=2023.1.0=hdb19cb5_46306
iopath=0.1.10=pypi_0
ipykernel=6.29.4=pypi_0
ipython=8.25.0=pypi_0
ipywidgets=8.1.3=pypi_0
isoduration=20.11.0=pypi_0
itsdangerous=2.2.0=pypi_0
jedi=0.19.1=pypi_0
jinja2=3.1.4=py311h06a4308_0
joblib=1.4.2=pypi_0
jpeg=9e=h5eee18b_1
json5=0.9.25=pypi_0
jsonpointer=2.4=pypi_0
jsonschema=4.22.0=pypi_0
jsonschema-specifications=2023.12.1=pypi_0
jupyter=1.0.0=pypi_0
jupyter-client=8.6.2=pypi_0
jupyter-console=6.6.3=pypi_0
jupyter-core=5.7.2=pypi_0
jupyter-events=0.10.0=pypi_0
jupyter-lsp=2.2.5=pypi_0
jupyter-server=2.14.1=pypi_0
jupyter-server-terminals=0.5.3=pypi_0
jupyterlab=4.2.1=pypi_0
jupyterlab-pygments=0.3.0=pypi_0
jupyterlab-server=2.27.2=pypi_0
jupyterlab-widgets=3.0.11=pypi_0
kiwisolver=1.4.5=pypi_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libcublas=11.11.3.6=0
libcufft=10.9.0.58=0
libcufile=1.10.0.4=0
libcurand=10.3.6.39=0
libcusolver=11.4.1.48=0
libcusparse=11.7.5.86=0
libdeflate=1.17=h5eee18b_1
libffi=3.4.4=h6a678d5_1
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h5eee18b_3
libidn2=2.3.4=h5eee18b_0
libjpeg-turbo=2.0.0=h9bf148f_0
libnpp=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libpng=1.6.39=h5eee18b_0
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.5.1=h6a678d5_0
libunistring=0.9.10=h27cfd23_0
libuuid=1.41.5=h5eee18b_0
libwebp-base=1.3.2=h5eee18b_0
lit=15.0.7=pypi_0
llvm-openmp=14.0.6=h9e868ea_0
loguru=0.7.2=pypi_0
lz4-c=1.9.4=h6a678d5_1
markupsafe=2.1.3=py311h5eee18b_0
matplotlib=3.9.0=pypi_0
matplotlib-inline=0.1.7=pypi_0
missingno=0.5.2=pypi_0
mistune=3.0.2=pypi_0
mkl=2023.1.0=h213fc3f_46344
mkl-service=2.4.0=py311h5eee18b_1
mkl_fft=1.3.8=py311h5eee18b_0
mkl_random=1.2.4=py311hdb19cb5_0
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpmath=1.3.0=py311h06a4308_0
multimethod=1.11.2=pypi_0
nbclient=0.10.0=pypi_0
nbconvert=7.16.4=pypi_0
nbformat=5.10.4=pypi_0
ncurses=6.4=h6a678d5_0
ndg-httpsclient=0.5.1=pypi_0
nest-asyncio=1.6.0=pypi_0
nettle=3.7.3=hbbd107a_1
networkx=3.1=py311h06a4308_0
notebook=7.2.0=pypi_0
notebook-shim=0.2.4=pypi_0
numpy=1.26.4=py311h08b1b3b_0
numpy-base=1.26.4=py311hf175353_0
objaverse=0.1.7=pypi_0
open3d=0.18.0=pypi_0
opencv-python=4.9.0.80=pypi_0
opencv-python-headless=4.8.0.74=pypi_0
openh264=2.1.1=h4ff587b_0
openjpeg=2.4.0=h3ad879b_0
openssl=3.0.13=h7f8727e_2
overrides=7.7.0=pypi_0
packaging=24.0=pypi_0
pandas=2.2.2=pypi_0
pandas-profiling=3.2.0=pypi_0
pandocfilters=1.5.1=pypi_0
parso=0.8.4=pypi_0
pexpect=4.9.0=pypi_0
phik=0.12.4=pypi_0
pillow=10.3.0=py311h5eee18b_0
pip=24.0=py311h06a4308_0
platformdirs=4.2.2=pypi_0
plotly=5.22.0=pypi_0
portalocker=2.8.2=pypi_0
prometheus-client=0.20.0=pypi_0
prompt-toolkit=3.0.45=pypi_0
proto-plus=1.23.0=pypi_0
protobuf=4.25.3=pypi_0
psutil=5.9.8=pypi_0
ptyprocess=0.7.0=pypi_0
pure-eval=0.2.2=pypi_0
pyarrow=16.1.0=pypi_0
pyasn1=0.6.0=pypi_0
pyasn1-modules=0.4.0=pypi_0
pycocotools=2.0.7=pypi_0
pycparser=2.22=pypi_0
pydantic=2.7.2=pypi_0
pydantic-core=2.18.3=pypi_0
pygments=2.18.0=pypi_0
pyopenssl=24.2.1=pypi_0
pyparsing=3.1.2=pypi_0
pyquaternion=0.9.9=pypi_0
pysocks=1.7.1=py311h06a4308_0
python=3.11.9=h955ad1f_0
python-dateutil=2.9.0.post0=pypi_0
python-dotenv=1.0.1=pypi_0
python-json-logger=2.0.7=pypi_0
python-magic=0.4.27=pypi_0
pytorch-cuda=11.8=h7e8668a_5
pytorch-mutex=1.0=cuda
pytorch3d=0.7.6=pypi_0
pytz=2024.1=pypi_0
pywavelets=1.6.0=pypi_0
pyyaml=6.0.1=py311h5eee18b_0
pyzmq=26.0.3=pypi_0
qtconsole=5.5.2=pypi_0
qtpy=2.4.1=pypi_0
readline=8.2=h5eee18b_0
referencing=0.35.1=pypi_0
regex=2024.5.15=pypi_0
requests=2.32.2=py311h06a4308_0
requests-toolbelt=1.0.0=pypi_0
retrying=1.3.4=pypi_0
rfc3339-validator=0.1.4=pypi_0
rfc3986-validator=0.1.1=pypi_0
roboflow=1.1.30=pypi_0
rpds-py=0.18.1=pypi_0
rsa=4.9=pypi_0
safetensors=0.4.3=pypi_0
scikit-learn=1.5.0=pypi_0
scipy=1.13.1=pypi_0
seaborn=0.13.2=pypi_0
segment-anything=1.0=pypi_0
send2trash=1.8.3=pypi_0
sentence-transformers=3.0.1=pypi_0
sentencepiece=0.2.0=pypi_0
setuptools=69.5.1=py311h06a4308_0
shapely=2.0.4=pypi_0
six=1.16.0=pypi_0
sniffio=1.3.1=pypi_0
soupsieve=2.5=pypi_0
sqlite=3.45.3=h5eee18b_0
stack-data=0.6.3=pypi_0
supervision=0.20.0=pypi_0
sympy=1.12=py311h06a4308_0
tabulate=0.9.0=pypi_0
tangled-up-in-unicode=0.2.0=pypi_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.3.0=pypi_0
termcolor=2.4.0=pypi_0
terminado=0.18.1=pypi_0
threadpoolctl=3.5.0=pypi_0
timm=1.0.3=pypi_0
tinycss2=1.3.0=pypi_0
tk=8.6.14=h39e8969_0
tokenizers=0.19.1=pypi_0
tomli=2.0.1=pypi_0
torch=2.0.1+cu118=pypi_0
torchaudio=2.1.0=py311_cu118
torchvision=0.15.2+cu118=pypi_0
tornado=6.4=pypi_0
tqdm=4.66.4=pypi_0
traitlets=5.14.3=pypi_0
transformers=4.41.2=pypi_0
trimesh=4.4.1=pypi_0
triton=2.0.0=pypi_0
types-python-dateutil=2.9.0.20240316=pypi_0
typing_extensions=4.11.0=py311h06a4308_0
tzdata=2024.1=pypi_0
uri-template=1.3.0=pypi_0
uritemplate=4.1.1=pypi_0
urllib3=2.2.1=py311h06a4308_0
visions=0.7.4=pypi_0
wcwidth=0.2.13=pypi_0
webcolors=1.13=pypi_0
webencodings=0.5.1=pypi_0
websocket-client=1.8.0=pypi_0
werkzeug=3.0.3=pypi_0
wheel=0.43.0=py311h06a4308_0
widgetsnbextension=4.0.11=pypi_0
xz=5.4.6=h5eee18b_1
yacs=0.1.8=pypi_0
yaml=0.2.5=h7b6447c_0
yapf=0.40.2=pypi_0
zipp=3.19.1=pypi_0
zlib=1.2.13=h5eee18b_1
zstandard=0.22.0=pypi_0
zstd=1.5.5=hc292b87_2
